fix issue where phrases are detected across terminating chars
